Production order crud done

+++ src/main/java/ma/youcode/supplyChainX/repository/ProductionOrderRepository.java
@@ -3,5 +3,5 @@ package ma.youcode.supplyChainX.repository;
 import ma.youcode.supplyChainX.model.ProductionOrder;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface ProductionOrderRepository extends JpaRepository<ProductionOrder, Integer> {
+public interface ProductionOrderRepository extends JpaRepository<ProductionOrder, Long> {
 }
